+++ Src/Geometry/2d/Point_abstr.cpp
@@ -19,7 +19,7 @@
 
 namespace Geom2d {
 
-bool point_within(glm::vec2 pt, glm::vec2 top_left, glm::vec2 bottom_right)
+bool point_within(glm::dvec2 pt, glm::dvec2 top_left, glm::dvec2 bottom_right)
 {
 	if(top_left.x <= pt.x && bottom_right.x >= pt.x && top_left.y >= pt.y && bottom_right.y <= pt.y)
 		return true;
