+++ Src/Geometry/2d/Point_abstr.hpp
@@ -27,7 +27,7 @@
 
 namespace Geom2d {
 
-bool point_within(glm::vec2 pt, glm::vec2 top_left, glm::vec2 bottom_right);
+bool point_within(glm::dvec2 pt, glm::dvec2 top_left, glm::dvec2 bottom_right);
 
 template <class Impl>
 class Point_abstr : public Geom2d_abstr<Point_abstr<Impl>> {
@@ -37,23 +37,23 @@ public:
 	{
 
 	}
-	glm::vec2 pos()
+	glm::dvec2 pos()
 	{
 		return static_cast<Impl*>(this)->pos_impl();
 	}
 
-	glm::vec2 at_geom(float t) { return pos(); }
-	bool within_geom(glm::vec2 top_left, glm::vec2 bottom_right, bool contained)
+	glm::dvec2 at_geom(double t) { return pos(); }
+	bool within_geom(glm::dvec2 top_left, glm::dvec2 bottom_right, bool contained)
 	{
 		return point_within(pos(), top_left, bottom_right);
 	}
-	float closest_to_point_interp_geom(glm::vec2 const& pt) { return 0.0f; }
-	glm::vec2 closest_to_point_geom(glm::vec2 const& pt) { return pos(); }
-	float dist_to_point_geom(glm::vec2 const& pt)
+	double closest_to_point_interp_geom(glm::dvec2 const& pt) { return 0.0f; }
+	glm::dvec2 closest_to_point_geom(glm::dvec2 const& pt) { return pos(); }
+	double dist_to_point_geom(glm::dvec2 const& pt)
 	{
 		return glm::distance(pos(), pt);
 	}
-	float dist_to_point_2(glm::vec2 const& pt)
+	double dist_to_point_2(glm::dvec2 const& pt)
 	{
 		return glm::distance2(pos(), pt);
 	}
@@ -63,7 +63,7 @@ class Point : public Point_abstr<Point> {
 private:
 	Param mX, mY;
 public:
-	Point(glm::vec2 pt)
+	Point(glm::dvec2 pt)
 		: mX(pt.x)
 		, mY(pt.y)
 	{
@@ -76,9 +76,9 @@ public:
 
 	}
 
-	glm::vec2 pos_impl()
+	glm::dvec2 pos_impl()
 	{
-		return glm::vec2(mX.val(), mY.val());
+		return glm::dvec2(mX.val(), mY.val());
 	}
 
 	Param* x() { return &mX; }
